+++ src/examples/program.ts
@@ -1,7 +1,7 @@
 import {
   SelfProof,
   Field,
-  Experimental,
+  ZkProgram,
   verify,
   isReady,
   Proof,
@@ -12,7 +12,7 @@ import {
 
 await isReady;
 
-let MyProgram = Experimental.ZkProgram({
+let MyProgram = ZkProgram({
   publicOutput: Field,
 
   methods: {
@@ -36,7 +36,7 @@ let MyProgram = Experimental.ZkProgram({
 MyProgram.publicInputType satisfies Provable<Empty>;
 MyProgram.publicOutputType satisfies typeof Field;
 
-let MyProof = Experimental.ZkProgram.Proof(MyProgram);
+let MyProof = ZkProgram.Proof(MyProgram);
 
 console.log('program digest', MyProgram.digest());