+++ src/tests/inductive-proofs.ts
@@ -1,7 +1,7 @@
 import {
   SelfProof,
   Field,
-  Experimental,
+  ZkProgram,
   isReady,
   shutdown,
   Proof,
@@ -10,7 +10,7 @@ import { tic, toc } from '../examples/zkapps/tictoc.js';
 
 await isReady;
 
-let MaxProofsVerifiedZero = Experimental.ZkProgram({
+let MaxProofsVerifiedZero = ZkProgram({
   publicInput: Field,
 
   methods: {
@@ -24,7 +24,7 @@ let MaxProofsVerifiedZero = Experimental.ZkProgram({
   },
 });
 
-let MaxProofsVerifiedOne = Experimental.ZkProgram({
+let MaxProofsVerifiedOne = ZkProgram({
   publicInput: Field,
 
   methods: {
@@ -47,7 +47,7 @@ let MaxProofsVerifiedOne = Experimental.ZkProgram({
   },
 });
 
-let MaxProofsVerifiedTwo = Experimental.ZkProgram({
+let MaxProofsVerifiedTwo = ZkProgram({
   publicInput: Field,
 
   methods: {
@@ -100,7 +100,7 @@ async function testRecursion(
 ) {
   console.log(`testing maxProofsVerified = ${maxProofsVerified}`);
 
-  let ProofClass = Experimental.ZkProgram.Proof(Program);
+  let ProofClass = ZkProgram.Proof(Program);
 
   tic('executing base case');
   let initialProof = await Program.baseCase(Field(0));